Rename PreferenceHandler and move to package org.dockfx
Also add some more Java docs
